+++ resolwe/flow/executors/null/run.py
@@ -22,7 +22,3 @@ class FlowExecutor(BaseFlowExecutor):  # pylint: disable=abstract-method
     def run(self, data_id, script, verbosity=1):
         """Do nothing :)."""
         send_manager_command(ExecutorProtocol.FINISH)
-
-    def terminate(self, data_id):
-        """Do nothing :)."""
-        pass